+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+### 1.0.6
+
+- Fixed issue where tasks didn't execute as provided
+- Added support for dual core PSoC5 project building (by using folder tags in the build file: eg: `(folder:source/CortexM4)`)
+- Added CMake support
+- Improved Clean / Build logic
+
+### 1.0.5
+
+- Various fixes (statusbar buttons etc.)
+
 ### 1.0.4
 
 - Various fixes (version updating, project check etc.)